refactor: Separate header components into styles and components
