+++ src/components/Header/Header.styles.ts
@@ -1,0 +1,12 @@
+export const headerStyle = {
+  padding: "20px 0",
+  lineHeight: "1.5em"
+};
+export const titleStyle = {
+  fontSize: "6rem",
+  fontWeight: "600",
+  marginBottom: "2rem",
+  lineHeight: "1em",
+  color: "#ececec",
+  TextAlign: "center"
+};
