+++ src/components/Header/index.tsx
@@ -1,0 +1,11 @@
+import * as S from "./Header.styles";
+
+const Header = () => {
+  return (
+    <header style={S.headerStyle}>
+      <h1 style={S.titleStyle}>Toodos</h1>
+    </header>
+  );
+};
+
+export default Header;
